✨ [feat] target 설정을 위한 뷰 분리
