+++ Projects/LastDance/LastDance/Sources/Presentation/Archive/Views/ArtReactionView.swift
@@ -197,106 +197,3 @@ struct ArtReactionView: View {
         }
     }
 }
-
-// MARK: - TabBarView Component
-struct TabBarView: View {
-    @Binding var selectedTab: ArtReactionTab
-    
-    var body: some View {
-        HStack(spacing: 0) {
-            Button(action: {
-                selectedTab = .artwork
-            }) {
-                VStack(spacing: 8) {
-                    Text("작품")
-                        .font(Font.custom("Pretendard", size: 18).weight(.semibold))
-                        .foregroundColor(selectedTab == .artwork ? LDColor.color1 : LDColor.color2)
-                    
-                    Rectangle()
-                        .fill(selectedTab == .artwork ? LDColor.color1 : Color.clear)
-                        .frame(height: 2)
-                }
-            }
-            .frame(maxWidth: .infinity)
-            
-            Button(action: {
-                selectedTab = .reaction
-            }) {
-                VStack(spacing: 8) {
-                    Text("감상")
-                        .font(Font.custom("Pretendard", size: 18).weight(.semibold))
-                        .foregroundColor(selectedTab == .reaction ? LDColor.color1 : LDColor.color2)
-                    
-                    Rectangle()
-                        .fill(selectedTab == .reaction ? LDColor.color1 : Color.clear)
-                        .frame(height: 2)
-                }
-            }
-            .frame(maxWidth: .infinity)
-        }
-        .padding(.horizontal, 20)
-    }
-}
-
-// MARK: - ScrollOffsetPreferenceKey
-struct ScrollOffsetPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = nextValue()
-    }
-}
-
-// MARK: - ScrollViewObserver
-struct ScrollViewObserver<Content: View>: UIViewRepresentable {
-    let content: Content
-    @Binding var scrollOffset: CGFloat
-    
-    init(scrollOffset: Binding<CGFloat>, @ViewBuilder content: () -> Content) {
-        _scrollOffset = scrollOffset
-        self.content = content()
-    }
-    func makeUIView(context: Context) -> UIScrollView {
-        let scrollView = UIScrollView()
-        scrollView.delegate = context.coordinator
-        scrollView.showsVerticalScrollIndicator = true
-        scrollView.showsHorizontalScrollIndicator = false
-        scrollView.bounces = true
-        
-        let hostingController = UIHostingController(rootView: content)
-        hostingController.view.translatesAutoresizingMaskIntoConstraints = false
-        
-        scrollView.addSubview(hostingController.view)
-        
-        NSLayoutConstraint.activate([
-            hostingController.view.topAnchor.constraint(equalTo: scrollView.topAnchor),
-            hostingController.view.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor),
-            hostingController.view.trailingAnchor.constraint(equalTo: scrollView.trailingAnchor),
-            hostingController.view.bottomAnchor.constraint(equalTo: scrollView.bottomAnchor),
-            hostingController.view.widthAnchor.constraint(equalTo: scrollView.widthAnchor)
-        ])
-        
-        context.coordinator.hostingController = hostingController
-        
-        return scrollView
-    }
-    func updateUIView(_ uiView: UIScrollView, context: Context) {
-        context.coordinator.hostingController?.rootView = content
-    }
-    func makeCoordinator() -> Coordinator {
-        Coordinator(scrollOffset: $scrollOffset)
-    }
-    class Coordinator: NSObject, UIScrollViewDelegate {
-        @Binding var scrollOffset: CGFloat
-        var hostingController: UIHostingController<Content>?
-        
-        init(scrollOffset: Binding<CGFloat>) {
-            _scrollOffset = scrollOffset
-        }
-        func scrollViewDidScroll(_ scrollView: UIScrollView) {
-            let offset = max(0, scrollView.contentOffset.y)
-            DispatchQueue.main.async {
-                self.scrollOffset = offset
-            }
-        }
-    }
-}
